requirements.txt
python-telegram-bot
fastapi
uvicorn
aiohttp
python-dotenv
python-multipart
pillow